update to MessageDeframer.MessageProducer implementation:
everything still runs in transport thread

tests pass (except for MessageDeframer tests, which used mocks and are not valid
for the new structure)

AbstractStream2.TransportState.MessagesAvailable() directly loops over the messages
and calls messageRead() - this was necessary to get the tests with mocks (in Netty,
at least) passing. It's possible non-mock tests were also broken but masked by this
change.